Refactor types; address comments
