+++ src/plugins/vis_augmenter/public/expressions/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './vis_layers';
+export * from './types';
