+++ src/plugins/vis_augmenter/public/index.ts
@@ -18,14 +18,8 @@ export {
   SavedObjectOpenSearchDashboardsServicesWithAugmentVis,
 } from './saved_augment_vis';
 
-export {
-  ISavedAugmentVis,
-  VisLayerExpressionFn,
-  AugmentVisSavedObject,
-  VisLayerFunctionDefinition,
-  VisLayer,
-  VisLayers,
-} from './types';
+export { VisLayer, VisLayers, VisLayerTypes } from './types';
 
 export * from './expressions';
 export * from './utils';
+export * from './saved_augment_vis';
