+++ src/plugins/vis_augmenter/public/utils/utils.ts
@@ -19,6 +19,11 @@ export const isEligibleForVisLayers = (vis: Vis): boolean => {
   return vis.params.type === 'line';
 };
 
+/**
+ * Using a SavedAugmentVisLoader, fetch all saved objects that are of 'augment-vis' type
+ * and filter out to return the ones associated to the particular vis via
+ * matching vis ID.
+ */
 export const getAugmentVisSavedObjs = async (
   visId: string | undefined,
   loader: SavedAugmentVisLoader | undefined
@@ -32,6 +37,13 @@ export const getAugmentVisSavedObjs = async (
   }
 };
 
+/**
+ * Given an array of augment-vis saved objects that contain expression function details,
+ * construct a pipeline that will execute each of these expression functions.
+ * Note that the order does not matter; each expression function should be taking
+ * in the current output and appending its results to it, such that the end result
+ * contains the results from each expression function that was ran.
+ */
 export const buildPipelineFromAugmentVisSavedObjs = (objs: ISavedAugmentVis[]): string => {
   const visLayerExpressionFns = [] as Array<
     ExpressionAstFunctionBuilder<VisLayerFunctionDefinition>
